pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prcat"
version = "0.1.0"
description = "AI-powered PR reviewer — understand any pull request in seconds"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["ai", "pull-request", "code-review", "git", "cli", "developer-tools", "llm", "github"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "anthropic>=0.20.0",
    "openai>=1.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
prcat = "prcat.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "click>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/prcat"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=prcat --cov-report=term-missing"
